+++ tests/Db/Expression/Function/ArrayMergeTest.php
@@ -25,4 +25,15 @@ final class ArrayMergeTest extends TestCase
         $this->assertSame($expression, $expression->type($intColumn));
         $this->assertSame($intColumn, $expression->getType());
     }
+
+    public function testOrdered(): void
+    {
+        $expression = new ArrayMerge();
+
+        $this->assertFalse($expression->getOrdered());
+        $this->assertSame($expression, $expression->ordered());
+        $this->assertTrue($expression->getOrdered());
+        $this->assertSame($expression, $expression->ordered(false));
+        $this->assertFalse($expression->getOrdered());
+    }
 }